magnetar/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["feus <barca105@nate.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
databases = {extras = ["postgresql"], version = "^0.5.5"}
python-dotenv = "^0.20.0"
asyncpg = "^0.25.0"
strawberry-graphql = "^0.113.0"
rich = "^12.4.4"
greenlet = "^1.1.2"
PyJWT = "^2.4.0"
google-auth = "^2.6.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
